customer_chat_analysis/requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
nltk
spacy
python-dotenv
google-generativeai
